test: split city credit and payment attempts tests
