+++ test/helpers/city-credit-fixtures/ledger-entries.ts
@@ -1,0 +1,5 @@
+export const REQUEST_ID = 'frontier-request-0001'
+
+export const SOURCE_KEY = 'founder-extra-payment-0001'
+
+export const CREATED_AT = '2026-08-22T12:00:00.000Z'
